BETA 202402.12 - New component / object and Sanity fix.
Player.gd: Added a bool to easily deactivate SanityComponent. Also added a print to show that Health is depleting because of low sanity to make this clearer.

ReadableComponent: Simple component to add readable documents. Can be attached to any CogitoObject.

Cogito_Keypad: New object type: enter a code to unlock doors or trigger other things with the included signal. Need a better spot in the Demo scene for this.