BETA 202402.14 - Quest System update
Implmented the new Quest system:
In order for this to work, you need to add /COGITO/QuestSystemPD/CogitoQuestManager.gd to the autoload.

I've added a small test area to the Demo scene to show 2 Quests, one with counter and one without.
Currently Quests ARE NOT BEING SAVED. Expect further updated.