More menu updates
Reshuffled the way menus are structured a bit:
- Moved all options into a single OptionsTabMenu scene.
- Main menu and Pause Menu are now just plain menus that contain the OptionsTabMenu scene. This makes it so any updates to OptionsTabMenu will automatically be in both the Pause Menu and the Main Menu.
- This should also make it clearer as to when to apply any changes made in the options menu (currently this is unchanged)

- Renamed some scenes and scripts - This might be a bit confusing so bear with me:
Since I've removed the old pause menu that was still based on the original EasyMenu addon was removed a few commits ago, I've finally decided to rename the TabMenu to PauseMenu, as that's what it was used for anyway.
I've tried to rename the PauseMenu in all Demo scenes as well, except in Demo_01 which currently seems to be broken on my end.