Added throwable dart
Throwable dart:
This is a very specific kind of wieldable:
The projectile it spawns is the same pickup item which has the wieldable item as it's item reference.
Instead of ammo, it reduces the item amount on the stack.
Could use some polish and still throws some errors, but should work overall.