Player HUD Style Changes
Added class name
Reordered variables
Added variable region
Moved setup of player and attributes to its own function - this is so that if the player object is changed the HUD can be reset to the new player.
Removed any children from the ui_attribute_area to clear any UI from any previous player.