Cleanup commit (PLEASE READ)
Started renaming files and folders to be more in line with the style guide and make more sense.
This might break a few references, I've tried my best to check them all and make sure they are working but I might've missed something.
If you find something broken, please create an issue or if it's an easy fix, push a commit.

Thanks!