Gamepad nav fix for grid inventories.
Fixed the gamepad navigation for the grid inventory by manually setting the focus neighbors on instantiation of slots.

One remaining cosmetic issue:
When an item is being moved with the gamepad, the grabbed item icon always first moves to the center of the first slot (instead of the currently focused slot). As soon as the focus is moved the grabbed item positions correctly, so it is updated correctly, just that initial "pick up" seems off.