Included shapecasting to the InteractionRaycast.gd along with highlighters to debug its behavior
This commit modifies the existing InteractionRaycast.gd and to a lesser extent the PlayerInteractionComponent.gd.

InteractionRaycast now uses a shapecast when the raycast doesn't hit a valid collider. If the shapecast hits multiple valid colliders it will use a 'hotspot' and judge distances between it and valid colliders to determine which object the player is more likely intending to pick up.

There is a known issue regarding the ShapeCast3D node reference being lost on scene change.