New: CogitoQuickslotContainer and CogitoQuickSlots (WIP)
Started work on the new quick slot system (which can/will replace the hotbar)
This is still a very early WIP, but using the mouse, you can already assign items to the new quick slots and trigger them by moving an item (it needs to be moving with the mouse cursor) and then left-clicking on the slot you want to assign. Currently no game pad support!

To give it a try, go to inventory_interface.gd in the Player_HUD scene and make sure "is_using_hotbar" is turned OFF.