Further refined despawning behavior
Now despawning is handled after inventory is closed by player for all cases where container is emptied out. However the instant despawn and container closure functionality is still kept for cases where it may be necessary with the improvement that now inventory window is closed as well.

Respawning lootable containers should behave as expected during respawn and not pop the contents on player's screen if he exited out of the container with anything other than interact key.