Inventory fixes (please read)
Small update to cogito inventories.
The inventory resource now has a "owner" property which gets set on opening the inventory GUI / interface.
When an item is "used" the owner is now referenced. This allows more control over what happens when items are used in specific contexts.

For example: In this update, wieldables are changed so that they can not be used/equipped while in an external inventory. They can also not be moved while they are being wielded. This solves an issue where players would be able to equip a wieldable that was inside an external inventory - and then close the inventory screen and walk away from the container while still having said wieldable equipped.