Translation Key resource
As a lot of strings in Cogito are passed through code, these need to be updated to be key based.
Translation keys aren't easily added to the POT file, so I instead created a custom resource for it.

TranslationKeyDict resource is a basic container for strings meant to be used to generate keys for the POT file.
These keys can be references in any code and translated with Godots built in _tr()_ method.